% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghe.utils-package.R
\docType{package}
\name{ghe.utils-package}
\alias{ghe.utils}
\alias{ghe.utils-package}
\title{ghe.utils: Utilities for Working with GitHub Enterprise}
\description{
Set the location of your company's GitHub Enterprise address, and seamlessly interact with a GitHub Enterprise Instance from the R console.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.danieldsjoberg.com/ghe.utils}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID}) [copyright holder]

}
\keyword{internal}
