% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_ghe_network_access.R
\name{has_ghe_network_access}
\alias{has_ghe_network_access}
\title{Check for GHE Network Access}
\usage{
has_ghe_network_access()
}
\value{
logical
}
\description{
Check for GHE Network Access
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
has_ghe_network_access()
\dontshow{\}) # examplesIf}
}
\author{
Daniel D. Sjoberg
}
