% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghe_token.R
\name{token}
\alias{token}
\alias{ghe_token}
\alias{create_ghe_token}
\title{Retrieve GH Enterprise Personal Access Token (PAT)}
\usage{
ghe_token()

create_ghe_token(scopes = c("repo", "user", "gist", "workflow"))
}
\arguments{
\item{scopes}{Character vector of token scopes, pre-selected in the web form.
Final choices are made in the GitHub form. Read more about GitHub API
scopes at
\url{https://docs.github.com/apps/building-oauth-apps/understanding-scopes-for-oauth-apps/}.}
}
\description{
Functions for setting up a GHE PAT (\code{create_ghe_token()}),
and accessing the saved PAT (\code{ghe_token()}).
}
