% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ghe.R
\name{use_ghe}
\alias{use_ghe}
\title{Connect a local repo with GitHub Enterprise}
\usage{
use_ghe(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[usethis:use_github]{usethis::use_github}}
  \describe{
    \item{\code{organisation}}{If supplied, the repo will be created under this
organisation, instead of the login associated with the GitHub token
discovered for this \code{host}. The user's role and the token's scopes must be
such that you have permission to create repositories in this
\code{organisation}.}
    \item{\code{private}}{If \code{TRUE}, creates a private repository.}
    \item{\code{visibility}}{Only relevant for organisation-owned repos associated with
certain GitHub Enterprise products. The special "internal" \code{visibility}
grants read permission to all organisation members, i.e. it's intermediate
between "private" and "public", within GHE. When specified, \code{visibility}
takes precedence over \code{private = TRUE/FALSE}.}
    \item{\code{protocol}}{One of "https" or "ssh"}
  }}
}
\description{
Function in a wrapper for \code{usethis::use_github()} with the GitHub Enterprise
instance.
}
\seealso{
Review \code{\link[usethis:use_github]{usethis::use_github()}} for details.
}
