% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ghe_release.R
\name{use_ghe_date_release}
\alias{use_ghe_date_release}
\title{Make GHE Release}
\usage{
use_ghe_date_release(name = "<release annotation>", append_date = TRUE)
}
\arguments{
\item{name}{Name of GHE release}

\item{append_date}{Logical indicating whether to append the date to the
release name.}
}
\description{
Make GHE Release
}
